requirements.txt
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
Flask==1.1.1
freezegun==0.3.15
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.4.4
pytest==5.4.3
python-dotenv==0.12.0
requests==2.22.0
selenium==3.141.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.1
wrapt==1.11.2
